# 8315932: runtime/InvocationTests spend a lot of time on dependency verification