@shipilev 2023-09-11
Any takers? @lmesnik, @tstuefe, maybe?