@tstuefe 2023-10-04
Sorry, was off on PTO.

Not sure it works as intended though. Setting MaxMetaspaceSize does not expedite class unloading, it just makes class loading fail when that size is reached. It have worked because setting MaxMetaspaceSize influences the metaspace GC threshold, so we do GCs earlier and clean up unused class loaders earlier. But that can be done directly with "MetaspaceSize=very small value, e.g. 1M".

Bad case would be the speedup you see came from the tests bailing out early because they hit and then hid Metaspace OOMs.