@shipilev 2023-10-06
I am redoing this fix in #16052.